app-requirements.txt
aiofiles==0.5.0
altair==4.1.0
aniso8601==7.0.0
argon2-cffi==20.1.0
astor==0.8.1
async-exit-stack==1.0.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.23
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.1.0
easydict==1.9
email-validator==1.1.2
entrypoints==0.3
fastapi==0.65.1
gitdb==4.0.7
GitPython==3.1.17
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.2
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==2.2.3
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numpy==1.18.2
opencv-python==4.5.2.52
orjson==3.5.2
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.0
ptyprocess==0.7.0
pyarrow==4.0.0
pycocotools==2.0.2
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-multipart==0.0.5
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
redis==3.5.3
requests==2.25.1
Rx==1.6.1
scikit-image==0.16.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.16.0
smmap==4.0.0
starlette==0.14.2
streamlit==0.82.0
tensorboardX==2.0
terminado==0.10.0
testpath==0.5.0
toml==0.10.2
toolz==0.11.1
torch==1.4.0
tornado==6.1
tqdm==4.43.0
traitlets==4.3.3
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.14.0
validators==0.18.2
watchdog==2.1.2
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
widgetsnbextension==3.5.1
zipp==3.4.1
